requirements/project.txt
alabaster==0.7.12
amqp==2.5.1
argon2-cffi==19.1.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
billiard==3.5.0.5
boto3==1.9.229
botocore==1.12.229
celery==4.2.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Collectfast==0.6.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.13
django-allauth==0.38.0
django-anymail==5.0
django-appconf==1.0.3
django-contrib-comments==1.9.1
django-cors-headers==3.1.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.5
django-filter==2.2.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-model-utils==3.1.2
django-redis==4.10.0
django-rest==0.0.1
django-rest-auth==0.9.5
django-storages==1.7.1
django-uuslug==1.1.9
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.15.2
drf-yasg==1.16.1
entrypoints==0.3
factory-boy==2.11.1
Faker==2.0.1
flake8==3.7.5
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
inflection==0.3.1
ipdb==0.11
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.4
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.17.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==5.4.1
pluggy==0.13.0
prompt-toolkit==2.0.9
psycopg2==2.7.4
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydevd==1.7.1
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==4.2.0
pytest-django==3.4.7
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==2.0.1
python3-openid==3.1.0
pytz==2018.9
raven==6.10.0
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.2
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.3.5
typing-extensions==3.7.4
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1.3
zipp==0.6.0
